requirements.txt
seaborn
pandas
numpy
matplotlib
statsmodel
itertools